+++ changelog.md
@@ -9,7 +9,10 @@
 * New: settings now can render the `Clear` button for text-based field
 * New: checkbox based options can be used for the option switch control
 * New: check for capability for postback for main plugin panels
-* New: added several more UI icons
+* New: rewritten most of the settings processing for better validation
+* New: `Sanitize` class has new method for `deep` processing
+* New: `Sanitize` class renames some methods with proper names
+* New: added several more UI icons to the icon font
 * Edit: improvements to the default styling for dashboard and tabs
 * Edit: improvements to base `Panel` object content file names
 * Edit: improvements to the `License` base class related to API updates
@@ -18,6 +21,7 @@
 * Edit: outdated actions prefixes replacing `d4p` with `dev4press`
 * Edit: various improvements to sanitization and escaping
 * Edit: built-in cacert.pem 20240311
+* Deprecated: several methods in the `Sanitize` class
 * Fix: some `Core` properties were private not protected
 * Fix: license check job initialized with filter not action